project/realtime_sentiment/requirement.txt
opencv-python
pydub
kafka-python
transformers
torch
whisper
fer
deepface
TikTokApi
numpy
imageio
ffmpeg-python
python-dotenv
requests
pytube
moviepy
fastapi
torchaudio